+++ src/action/action.ts
@@ -1,10 +1,11 @@
 import { IResolvedOperation } from '../configuration/types';
+import { Diagnostics } from '../diagnostics';
 import { IEventbus } from '../eventbus/types';
 import { deepCopy } from '../operation/helper/deep-copy';
 import {
   IOperationContext,
   TOperationData,
-  TOperationResult
+  TOperationResult,
 } from '../operation/types';
 import { isPromise } from './is-promise';
 import { IAction } from './types';
@@ -19,10 +20,17 @@ export class Action implements IAction {
   ) {}
 
   start(initOperationData?: TOperationData): Promise<TOperationData> {
+    Diagnostics.active &&
+      Diagnostics.send(
+        'eligius-diagnostics-action',
+        `${this.name ?? 'Action'} begins executing start operations`
+      );
+
     const context: IOperationContext = {
       currentIndex: -1,
       eventbus: this.eventbus,
     };
+
     const result = new Promise<TOperationData>((resolve, reject) => {
       this.executeOperation(
         this.startOperations,
@@ -33,11 +41,24 @@ export class Action implements IAction {
         context
       );
     }).catch((e) => {
+      Diagnostics.active &&
+        Diagnostics.send('eligius-diagnostics-action-error', {
+          name,
+          error: e,
+        });
       console.error(`Error in action start '${this.name}'`);
       console.error(e);
       throw e;
     });
-    return result;
+    return !Diagnostics.active
+      ? result
+      : result.then((operationsResult) => {
+          Diagnostics.send(
+            'eligius-diagnostics-action',
+            `${this.name ?? 'Action'} finished executing start operations`
+          );
+          return operationsResult;
+        });
   }
 
   executeOperation(
@@ -77,6 +98,13 @@ export class Action implements IAction {
         ? deepCopy(operationInfo.operationData)
         : {};
       const mergedOperationData = Object.assign(previousOperationData, copy);
+
+      Diagnostics.active &&
+        Diagnostics.send('eligius-diagnostics-operation', {
+          systemName: operationInfo.systemName,
+          operationData: operationInfo.operationData,
+          context,
+        });
 
       const operationResult: TOperationResult = operationInfo.instance.call(
         context,