+++ src/diagnostics/diagnostics.ts
@@ -1,5 +1,8 @@
 import { TDiagnosticType } from './types';
 
 export class Diagnostics {
-  static send(_name: TDiagnosticType, _data: any): void {}
+  static send(_name: TDiagnosticType, _data: any): void {
+    console.log(_name, _data);
+  }
+  static active: boolean;
 }