+++ src/operation/helper/deep-copy.ts
@@ -1,11 +1,13 @@
 import { isDefined } from 'ts-is-present';
 
+const copyFunction =
+  typeof structuredClone !== 'undefined'
+    ? structuredClone
+    : (original: any) => JSON.parse(JSON.stringify(original));
+
 export function deepCopy<T>(original: T): T {
   if (isDefined(original)) {
-    if (typeof structuredClone !== 'undefined') {
-      return structuredClone<T>(original);
-    }
-    return JSON.parse(JSON.stringify(original)) as T;
+    return copyFunction(original) as T;
   }
   return original;
 }